+++ AzureIntegration/AzureCloud.cs
@@ -259,10 +259,7 @@ public class AzureCloud
 
         MSBuildLocator.RegisterDefaults();
 
-        // PublishUsingMsBuild1(projectFile, publishDirectory);
-        // PublishUsingMsBuild2(projectFile, publishDirectory);
-        // PublishUsingMsBuild3(projectFile, publishDirectory);
-        PublishUsingMsBuild4(projectFile, publishDirectory);
+        PublishProjectUsingMsBuild(projectFile, publishDirectory);
         return new DirectoryInfo(publishDirectory);
     }
 
@@ -279,84 +276,7 @@ public class AzureCloud
         return publishDirectory;
     }
 
-    private void PublishUsingMsBuild1(FileInfo projectFile, string publishDirectory)
-    {
-        var projectCollection = new ProjectCollection();
-        var buildRequestData = new BuildRequestData(projectFile.FullName, new Dictionary<string, string>
-        {
-            { "Configuration", "Release" },
-            { "OutputPath", publishDirectory }
-        }, null, ["Publish"], null);
-
-        var buildParameters = new BuildParameters(projectCollection)
-        {
-            Loggers = [new ConsoleLogger(LoggerVerbosity.Minimal)]
-        };
-
-        var buildResult = BuildManager.DefaultBuildManager.Build(buildParameters, buildRequestData);
-        if (buildResult.OverallResult == BuildResultCode.Failure)
-        {
-            throw new InvalidOperationException("Project publish failed");
-        }
-
-        Console.WriteLine("Project published to: {0}", publishDirectory);
-    }
-
-    private void PublishUsingMsBuild2(FileInfo projectFile, string publishDirectory)
-    {
-        var project = ProjectRootElement.Open(projectFile.FullName);
-        var projectInstance = new ProjectInstance(project);
-
-        var publishTarget = projectInstance.Targets.FirstOrDefault(t => t.Value.Name == "Publish").Value;
-        if (publishTarget == null)
-        {
-            throw new InvalidOperationException("Publish target not found in project file");
-        }
-
-        var buildRequestData = new BuildRequestData(projectInstance, ["Publish"]);
-        var buildParameters = new BuildParameters(ProjectCollection.GlobalProjectCollection)
-        {
-            Loggers = [new ConsoleLogger(LoggerVerbosity.Minimal)]
-        };
-
-        var buildResult = BuildManager.DefaultBuildManager.Build(buildParameters, buildRequestData);
-        if (buildResult.OverallResult == BuildResultCode.Failure)
-        {
-            throw new InvalidOperationException("Project publish failed");
-        }
-
-        Console.WriteLine("Project published to: {0}", publishDirectory);
-    }
-
-    private void PublishUsingMsBuild3(FileInfo projectFile, string publishDirectory)
-    {
-        var globalProperties = new Dictionary<string, string>
-        {
-            { "Configuration", "Release" },
-            { "OutputPath", publishDirectory }
-        };
-
-        var projectCollection = new ProjectCollection();
-        var buildRequestData = new BuildRequestData(projectFile.FullName, globalProperties, null, ["Publish"], null);
-
-        var buildParameters = new BuildParameters(projectCollection)
-        {
-            Loggers = null // You can add custom loggers if needed
-        };
-
-        var buildResult = BuildManager.DefaultBuildManager.Build(buildParameters, buildRequestData);
-
-        if (buildResult.OverallResult == BuildResultCode.Success)
-        {
-            Console.WriteLine("Build and publish succeeded!");
-        }
-        else
-        {
-            Console.WriteLine("Build and publish failed.");
-        }
-    }
-
-    private void PublishUsingMsBuild4(FileInfo projectFile, string publishDirectory)
+    private void PublishProjectUsingMsBuild(FileInfo projectFile, string publishDirectory)
     {
         Console.WriteLine($"Publishing {projectFile.Name} to {publishDirectory}");
         var globalProperties = new Dictionary<string, string>
@@ -400,125 +320,5 @@ public class AzureCloud
             resourceGroups.Add(new ResourceGroup(resourceGroup));
         }
         return resourceGroups;
-    }
-}
-
-public record ResourceGroup(ResourceGroupResource resourceGroup)
-{
-    public string Name => resourceGroup.Data.Name;
-    public async Task<StorageAccount> CreateStorageAccount(string name){
-        var storageAccountName = $"{name.ToLower().Replace("-", "")}storage";
-        if (storageAccountName.Length > 24)
-        {
-            storageAccountName = storageAccountName.Substring(0, 24);
-        }
-        var storageSku = new StorageSku(StorageSkuName.StandardLrs);
-        var storageKind = StorageKind.StorageV2;
-        var storageParameters = new StorageAccountCreateOrUpdateContent(storageSku, storageKind, location: resourceGroup.Data.Location);
-
-        var storageAccount = await resourceGroup.GetStorageAccounts().CreateOrUpdateAsync(
-            WaitUntil.Completed, storageAccountName, storageParameters);
-
-        return new StorageAccount(storageAccount.Value);
-    }
-
-    public async Task<AppServicePlan> CreateAppServicePlan(string name){
-        var appServicePlanData = new AppServicePlanData(resourceGroup.Data.Location)
-        {
-            Sku = new AppServiceSkuDescription { Name = "Y1",Tier = "Dynamic" },
-            Kind = "FunctionApp",            
-            IsReserved = true //Use linux
-        };
-
-        var appServicePlan = await resourceGroup.GetAppServicePlans().CreateOrUpdateAsync(
-            WaitUntil.Completed, name, appServicePlanData);
-
-        return new AppServicePlan(appServicePlan.Value);
-    }
-}
-
-public class StorageAccount
-{
-    private readonly StorageAccountResource _storageAccount;
-    private string _connectionString;
-
-    public StorageAccount(StorageAccountResource storageAccount)
-    {
-        _storageAccount = storageAccount;
-    }
-
-    public string Name => _storageAccount.Data.Name;
-    public string ConnectionString
-    {
-        get
-        {
-            return _connectionString ??= GetConnectionStringAsync();
-
-        }
-    }
-
-    public string GetConnectionStringAsync()
-    {
-        var keys = _storageAccount.GetKeysAsync().ToBlockingEnumerable();
-        var key = keys.First().Value;
-        return $"DefaultEndpointsProtocol=https;AccountName={Name};AccountKey={key};EndpointSuffix=core.windows.net";
-    }
-    
-}
-
-public class AppServicePlan
-{
-    private readonly AppServicePlanResource _appServicePlan;
-
-    public AppServicePlan(AppServicePlanResource appServicePlan)
-    {
-        _appServicePlan = appServicePlan;
-    }
-
-    public string Name => _appServicePlan.Data.Name;
-    public ResourceIdentifier Id => _appServicePlan.Id;
-}
-
-public class AzureFunction : IAsyncDisposable
-{
-    private readonly WebSiteResource _functionApp;
-    private readonly string _resourceGroupName;
-    private readonly AzureCloud _azureCloud;
-
-    public AzureFunction(WebSiteResource functionApp, string resourceGroupName, AzureCloud azureCloud)
-    {
-        _functionApp = functionApp;
-        _resourceGroupName = resourceGroupName;
-        _azureCloud = azureCloud;
-    }
-
-    public string Name => _functionApp.Data.Name;
-    public string Url => _functionApp.Data.DefaultHostName;
-
-    public async ValueTask DisposeAsync()
-    {
-        await _azureCloud.DeleteResourceGroup(_resourceGroupName);
-    }
-}
-
-public class AzureWebApp : IAsyncDisposable
-{
-    private readonly WebSiteResource _webApp;
-    private readonly string _resourceGroupName;
-    private readonly AzureCloud _azureCloud;
-
-    public AzureWebApp(WebSiteResource webApp, string resourceGroupName, AzureCloud azureCloud)
-    {
-        _webApp = webApp;
-        _resourceGroupName = resourceGroupName;
-        _azureCloud = azureCloud;
-    }
-
-    public string Name => _webApp.Data.Name;
-    public string Url => _webApp.Data.DefaultHostName;
-
-    public async ValueTask DisposeAsync()
-    {
-        await _azureCloud.DeleteResourceGroup(_resourceGroupName);
     }
 }
